feat(page): 작성자만 수정 삭제 버튼 활성화 & 삭제시 confirm

+++ src/main/resources/static/js/posts/detail.js
@@ -1,7 +1,25 @@
 document.addEventListener("DOMContentLoaded", function() {
     const postId = window.location.pathname.split("/").pop(); // URL 에서 ID를 추출
+    let userId = 0;
+    let userIdByPost = 0;
     fetchPostDetails(postId);
     fetchComments();
+
+    // 사용자 정보를 요청하여 userId를 가져옴
+    fetch("/api/me", {
+        method: "GET",
+        headers: {
+            'Content-Type': 'application/json'
+        },
+        credentials: 'include'  // 세션 쿠키를 포함하여 요청
+    })
+        .then(response => response.json())
+        .then(data => {
+            userId = data.id;
+        })
+        .catch(error => {
+            console.error('Error fetching user data:', error);
+        });
 
     // 게시글 상세 불러오기
     function  fetchPostDetails(id) {
@@ -33,7 +51,15 @@ document.addEventListener("DOMContentLoaded", function() {
 
         const date = new Date(post.createdAt);
         document.getElementById("post-createdAt").textContent = date.toISOString().slice(0, 16).replace('T', ' ');
+
+        // 현재 세션에서 가져온 userId와 post에서 가져온 userId가 같으면 수정삭제 버튼 활성화
+        if(post.userId === userId && post.userId !== 0) {
+            document.querySelector(".button-container").style.display = "block";
+        }
+        console.log("postUI:"+ post.userId)
+        console.log("userId: "+ userId);
     }
+
 
     // 수정 및 삭제
     const edit = document.getElementById("edit");
@@ -47,52 +73,36 @@ document.addEventListener("DOMContentLoaded", function() {
     // 삭제
     deletePost.addEventListener("click",function(e) {
         e.preventDefault()
+        const result = confirm("게시글을 삭제하시겠습니까?");
 
-        fetch(`/api/posts/${postId}`, {
-            method: "DELETE",
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            credentials: 'include' // 세션 포함
-        })
-            .then(response => {
-                if (!response.ok) {
-                    throw new Error(`HTTP error! status: ${response.status}`);
-                }
+        if(result) {
+            fetch(`/api/posts/${postId}`, {
+                method: "DELETE",
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                credentials: 'include' // 세션 포함
+            })
+                .then(response => {
+                    if (!response.ok) {
+                        throw new Error(`HTTP error! status: ${response.status}`);
+                    }
 
-                // 응답 본문이 있을 경우에만 JSON 파싱
-                return response.text().then(text => {
-                    return text ? JSON.parse(text) : {};
+                    // 응답 본문이 있을 경우에만 JSON 파싱
+                    return response.text().then(text => {
+                        return text ? JSON.parse(text) : {};
+                    });
+                })
+                .then(data => {
+                    console.log("Success:", data);
+                    window.location.href = "/posts";
+                })
+                .catch(error => {
+                    console.error("Error:", error);
                 });
-            })
-            .then(data => {
-                console.log("Success:", data);
-                window.location.href = "/posts";
-            })
-            .catch(error => {
-                console.error("Error:", error);
-            });
-
+        }
     });
 
-
-    // 사용자 정보를 요청하여 userId를 가져옴
-    let userId = null;
-
-    fetch("/api/me", {
-        method: "GET",
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        credentials: 'include'  // 세션 쿠키를 포함하여 요청
-    })
-        .then(response => response.json())
-        .then(data => {
-            userId = data.id;
-        })
-        .catch(error => {
-            console.error('Error fetching user data:', error);
-        });
 
     // 댓글 생성 api
     document.getElementById("input-comment").addEventListener("submit", function(event){
@@ -159,12 +169,22 @@ document.addEventListener("DOMContentLoaded", function() {
 
             const date = new Date(comment.createdAt).toISOString().slice(2, 16).replace("T"," ");
 
+            let display = "";
+            // 현재 세션에서 가져온 userId와 comment 에서 가져온 userId가 같으면 수정삭제 버튼 활성화
+            if(comment.userId === userId) {
+                display = "block";
+            } else {
+                display = "none";
+            }
+
             commentContainer.innerHTML = `
                 <span class="comment-content">${comment.content}</span>
                 <div class="comment-info">
                 <span><strong>${comment.nickname}</strong></span>
                 <span>${date}</span>
-                <a class="delete-comment" data-comment-id="${comment.id}"><img src="/images/x-mark.png" alt="댓글삭제버튼"></a>
+                <a class="delete-comment" data-comment-id="${comment.id}" style="display: ${display}">
+                    <img src="/images/x-mark.png" alt="댓글삭제버튼">
+                </a>
                 </div>            
             `;
 
@@ -182,22 +202,27 @@ document.addEventListener("DOMContentLoaded", function() {
     }
 
 
+
     // 댓글 삭제
     function deleteComment(commentId) {
-        fetch(`/api/posts/${postId}/comments/${commentId}`, {
-            method: "DELETE",
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            credentials: 'include' // 세션 포함
-        })
-            .then(response => {
-                if (!response.ok) {
-                    throw new Error(`HTTP error! status: ${response.status}`);
-                }
-                // 삭제 후 댓글 목록 갱신
-                fetchComments(); // 댓글 목록을 다시 불러오는 함수를 호출
+        const result = confirm("삭제하시겠습니까?");
+
+        if(result) {
+            fetch(`/api/posts/${postId}/comments/${commentId}`, {
+                method: "DELETE",
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                credentials: 'include' // 세션 포함
             })
-            .catch(error => console.error("Error deleting comment:", error));
+                .then(response => {
+                    if (!response.ok) {
+                        throw new Error(`HTTP error! status: ${response.status}`);
+                    }
+                    // 삭제 후 댓글 목록 갱신
+                    fetchComments(); // 댓글 목록을 다시 불러오는 함수를 호출
+                })
+                .catch(error => console.error("Error deleting comment:", error));
+        }
     }
 });
